Moved non-hooks out of hooks folder
