+++ react/src/utils/resolveAssembly.ts
